updated working app

+++ app/routes/api.py
@@ -27,7 +27,7 @@ def handle_integrity_error(err: IntegrityError):
 async def check_user():
     if session.get("user"):
         return AUTH_SUCCESS, 200
-    elif len(request.json) == 3:
+    elif request.json and len(request.json) == 3:
         password: str = request.json["password"]
         user = await User.get_or_none(
             user_name=request.json["user_name"],
@@ -42,15 +42,15 @@ async def check_user():
             session["user"] = request.json["user_name"]
             return {**SUCCESS, **AUTH_SUCCESS}, 200
         else:
-            return {**FAILED, **AUTH_FAILURE}, 400
+            return {**FAILED, **AUTH_FAILURE}, 200
     else:
-        return AUTH_FAILURE, 400
+        return AUTH_FAILURE, 200
 
 
 @api_router.post("/user")
 @execute_db
 async def create_user():
-    if len(request.json) == 4:
+    if request.json and len(request.json) == 4 and session.get("user") is None:
         password: str = request.json["password"]
         await User.create(
             user_name=request.json["user_name"],
@@ -70,7 +70,7 @@ async def create_user():
 @api_router.delete("/user")
 @execute_db
 async def delete_user():
-    if len(request.json) == 1 and "id" in request.json:
+    if request.json and len(request.json) == 1 and "id" in request.json:
         user: User = await User.get_or_none(id=request.json["id"])
 
         return SUCCESS if user and await user.delete() is None else FAILED
